Share specs between EmbeddedAnsible and AnsibleTower
