+++ spec/models/manageiq/providers/embedded_ansible/automation_manager/configuration_script_source_spec.rb
@@ -1,5 +1,5 @@
 require 'support/ansible_shared/automation_manager/configuration_script_source'
 
-describe ManageIQ::Providers::AnsibleTower::AutomationManager::ConfigurationScriptSource do
+describe ManageIQ::Providers::EmbeddedAnsible::AutomationManager::ConfigurationScriptSource do
   it_behaves_like 'ansible configuration_script_source'
 end
