+++ spec/models/manageiq/providers/embedded_ansible/automation_manager/credential_spec.rb
@@ -1,5 +1,5 @@
 require 'support/ansible_shared/automation_manager/credential'
 
-describe ManageIQ::Providers::AnsibleTower::AutomationManager::Credential do
+describe ManageIQ::Providers::EmbeddedAnsible::AutomationManager::Credential do
   it_behaves_like 'ansible credential'
 end
